GS-4637 The current product zip includes the following Hibernate jars under lib/Hibernate will be removed (ejb3-persistence.jar, hibernate3.jar, hibernate-commons-annotations.jar, hibernate-annotations.jar).
We will instruct 6.5 users to download the required Hibernate jar files using Maven template we will provide or asking them to download them by themselves if they don't use them already in other projects.
The purpose is not to maintain and deliver 3rd party jar files which are not by our runtime. In addition its important to note that the current GS zip did not contain several jar files synch as dom3j.jar which is a must for Hibernate and therefore required users to add it manually.

Reviewer: Shay B, Alex

svn path=/trunk/gigaspaces/; revision=26957


Former-commit-id: 31909543c53d9e5ba11220eef246b4ac61319a65