GS-5010: Configuration of disk-space monitor doesn't apply.
The configuration of disk-space monitor in services.config isn't applied - there is no possibility to enable it.

Bug: The component used to extract the 'enabled' property was wrong.
Fix: Encapsulated 'enabled' property into DiskSpace.isEnabled().

note - The Windows2k check was encapsulated as part of the DiskSpace construction.
A FlatlineMonitor will be used if the disk-space monitor is enabled, unless configured otherwise.

svn path=/trunk/gigaspaces/; revision=28200


Former-commit-id: e53e3f4eeb60ae6802e9153adbd447fe22069d75