GS-5058: Log formatter infers the caller's class and method names each time.
When calling the logger log methods, it infers the caller's class and method name, by parsing the stack trace of the call.
This is expensive and unnecessary when the log is configured to 'simple' format and not 'full' format.

A user can customize gs_logging to show either class/method name.

Reviewer: Kimchy

svn path=/trunk/gigaspaces/; revision=28849


Former-commit-id: 35fa66350e9673eeb58d7b3777a256a63b275d21