GS-5161: Services are not distributed evenly after redeployment on a larger set of GSCs.
The GSM manages an in-process counter for each InstantiatorResource (GSC). It increments the counter when it provisions, and decrements it when it deploys. But, it does not remove the in-process element when the decrement reaches zero. Thus, calls to InstantiatorResource.getInProcessCounter(..) will never return zero, but will return the total number of requested-deployments.

The GSM chooses a GSC, using the LeastActiveSelector which sorts them according to actual+pending (in process). The pending count is the result of InstantiatorResource.getInProcessCounter(..) - Unfortunately, the counter is always increasing, causing an empty GSC to act as if it was holding *previous* deployments!

svn path=/trunk/gigaspaces/; revision=29172


Former-commit-id: a5464547eb8c41d32189b7ccd9b0e044d9eb3b67