GS-5185
Allow beans in the application context to receive events when a space changes modes.

Currently it is done by beans implementing one of the interfaces: SpaceAfterBackupListener, SpaceAfterPrimaryListener, SpaceBeforeBackupListener, SpaceBeforePrimaryListener.
When a bean that implements one or more of these interfaces is defined in the application context it will receive events notifying it about spaces changing their mode.

Reviewed by kimchy.

svn path=/trunk/openspaces/; revision=29537


Former-commit-id: 096758f83b1759e504140c8480a0ddeceb3d091a