GS-5502: re-provision a service which failed to instantiate.
Delays next attempts to re-provision a service after it failed to allocate it on any of the existing GSCs. If a new GSC is brought up, a re-provision request will attempt immediately. Otherwise, after the specified delay, it will retry to provision the service on all existing GSCs. Default is 10 min.

svn path=/trunk/gigaspaces/; revision=38839


Former-commit-id: 41889029a0365a95657795389899bcfb5449916f