GS-5585: Redeployment is delayed although a new GSC exists.
Try to acquire the targeted GSC before dispatching a provision request. This is done to avoid provisioning of a service which will breach the required SLA.

If two GSCs load simultaneously, and there are two pending services (for example two backups with max-per-vm SLA), then it may happen that the first GSC to load will get a provision request for both. Instead, this fix will try to acquire the GSC, checking if it meets the SLA constraint, before deciding to provision it. Pending requests will be examined by the next GSC to load or at the next provision cycle.

svn path=/trunk/gigaspaces/; revision=39936


Former-commit-id: 93cd71f2fccc576fe60ada4ce46dbf3ae61a520d