GS-6106: One dedicated credential file for each single-space/cluster.
Each cluster can have it's own override file containing a specific url to the security file.

Example:

security1.properties:
    space-config.filters.DefaultSecurityFilter.enabled=true
    space-config.filters.DefaultSecurityFilter.url=security-file-1

security2.propterties:
    space-config.filters.DefaultSecurityFilter.enabled=true
    space-config.filters.DefaultSecurityFilter.url=security-file-2

Deploy the first cluster using the security1 properties file, and the second with the security2 properties file accordingly.

*Note* that the "com.gs.security.file.name" system property no longer influences the value of the url. This system property only serves the Service Grid components (GSC/GSM). This is opposed to the "com.gs.security.file.dir" system property (for now...)

If you want all the components to share the *same* file (spaces/clusters/GSC/GMS), you may set the url value to be resolved at deployment time.

Example:
    space-config.filters.DefaultSecurityFilter.enabled=true
    space-config.filters.DefaultSecurityFilter.url=${com.gs.security.file.name}
    
When a space member is initialized (as standalone or in a GSC), it will resolve the url to the value of "com.gs.security.file.name" system property.

svn path=/trunk/gigaspaces/; revision=45909


Former-commit-id: 62eb550236bb487fcb4c997ebd33c13074122677