GS-6203: Space takes a long time to load when loading a space outside of the gigaspaces root directory.
ResourceLoader.getResourceURL(urlPath, locatorBaseDir) recursively parses all the directory structure if the file isn't found in the classpath and isn't in the file-system. After the security enhancement (GS-6025: Root directory of security file URL is relative to com.gs.home and can't be modified.) the locatorBaseDir that is passed is no longer null. If the security file is no where to be found and the directory structure is deep, it will take a long time to parse, delaying the startup time of the space.

Added a boolean parameter 'derivePath' if true, and the resource could not be located in the file system using the urlPath and locatorBaseDir, we will try to recursively search under the locatorBaseDir as the base search path. By default, false.

svn path=/trunk/gigaspaces/; revision=46453


Former-commit-id: ffe2c3601d9390a6c28ae5af2ca8355ec235af87