GS-6361: Even distribution of primaries is no longer even.
After change GS-6278, Service instantiation is no longer done as part of advertise method. In order to extract the proxy we rely on the ServiceBeanInstance being returned. Since instantiation has moved to start method, the ServiceBeanInstance returns null before it had a chance to fully instantiate. If it's null, we can't query of it's active-election state and the algorithm decision is compromised.

Fix - Wait for instance to be instantiated or until it fails.

svn path=/trunk/gigaspaces/; revision=49246


Former-commit-id: ce33e598abcfc64205d2a65940f3ab64cc128a08