GS-6369: GSM failover
If there is no peer GSM that is managing an operation string, the recovering GSM will take control and activate the operational string that was lost. This healing mechanism will step in when the backup GSM failed to become primary for this operation string. This usually happens when the GSM failed to monitor it's primary. 

svn path=/trunk/gigaspaces/; revision=50271


Former-commit-id: 6c510acef5e0ac2203d186d1466fc01658d151eb