GS-6428: Secured GSC Proxy
A secured GSC creates a SecuredGSCProxy. For GSC-to-GSC invocations, this proxy uses a TrustedProxySecurityInterceptor. Clients acquiring this proxy, gain an UntrustedProxySecurityInterceptor. Upon authentication, a PrivilegedProxySecurityInterceptor used. Method invocations will be intercepted by these interceptors, each verifying the access level required vs. the one granted.

The CybernodeAdminProxy is wrapped by a SecuredCybernodeAdminProxy, constructed with the same security interceptor as for SecuredGSCProxy.

GS-6433: Security isn't enforced for destructive operations exposed by JConsole MBeans

For a secured GSC, a SecuredGSCExtension MBean is registered. Operations invoked on this MBean are intercepted, applying access restrictions based on the privileges granted upon authentication.

With a secured MBean, an authentication operation is exposed, allowing the user to authenticate before invoking any privileged operations.

svn path=/trunk/gigaspaces/; revision=51431


Former-commit-id: da28ffe910966667430bab73c48838454aa6e8bb