GS-6428: Secured GSA Proxy
A secured GSA creates a SecuredGSAProxy. For GSA-to-GSA invocations, this proxy uses a TrustedProxySecurityInterceptor. Clients acquiring this proxy, gain an UntrustedProxySecurityInterceptor. Upon authentication, a PrivilegedProxySecurityInterceptor used. Method invocations will be intercepted by these interceptors, each verifying the access level required vs. the one granted.

The ServiceAdmin is wrapped by a SecuredGSAAdminProxy, constructed with the same security interceptor as for SecuredGSAProxy.

GS-6433: Security isn't enforced for destructive operations exposed by JConsole MBeans

For a secured GSA, a SecuredGSAExtension MBean is registered. Operations invoked on this MBean are intercepted, applying access restrictions based on the privileges granted upon authentication.

With a secured MBean, an authentication operation is exposed, allowing the user to authenticate before invoking any privileged operations.

svn path=/trunk/gigaspaces/; revision=51579


Former-commit-id: f58bdb19abdde60e9c58feeeb640e88d5ccc2f6e