GS-6488
Exception caught at a processing unit initialization or disposal is first logged at the executing app domain before rethrowing it back to the root app domain.
This is done in order to keep a log of exceptions that for some reason are not serializable or contain data which is not serializable.

svn path=/trunk/gigaspaces-dotnet/; revision=51914


Former-commit-id: 26986f09a9ca0e0dd8a91fba17988e6261e8c5e4