GS-6504
[Class Loader Leak]: Bean Utils for POJO introspection holds a static cache of PojoProperties per class

It seems like without the cache, some tests fail. The reason (I think) for that is the fact that we have a bazillion gs.xml files within the QA, where evety time a proxy is created it needs to read them. This means that it need to reintrospect them again (since there is no cache). This slows down the space from starting up, which causes some tests to fail.

Surprisingly, some of the tests are not time based, so we might have problems with spaces that start up slower... . This is very important since we might have a case where a space does a recovery from a database, which means that it will start up slowly as well....


svn path=/trunk/gigaspaces/; revision=52130


Former-commit-id: 24987df919bc5cf72d7ce159c8750f20cdc51e08