GS-6560: Enhanced space security model (server and client side).
Ease configuration of the security manager by allowing the space properties file to include 
a property pointing to the security properties file.
com.gigaspaces.security.properties = my-file.properties

Allow the space properties file to point to itself ("this"), and include other properties related to security.
This eases the configuration by not having to manage a separate file also for security.

svn path=/trunk/gigaspaces/; revision=53531


Former-commit-id: f5dac8b81ba6b57b4a1a76bb69461f980b2d4823