GS-6794
As of now, last confirmed packets is not part of the redo log recovery process of a backup, as a result if a mirror is down and a backup is down and the backup comes back to life, it will think that last confirmed packet between the primary and the mirror is 0. If it becomes primary due to primary failure it will keep sending the same packet over and over once the mirror is up and the primary fails and the backup replaced it. this occurs because the mirror will keep telling him the last packet he received is 0 because the session was started with 0 from the backup.
Add special treatment to the case where the current mirror instance didn't receive any previous packets. 

(Bad English)

svn path=/trunk/gigaspaces/; revision=54996


Former-commit-id: c0017607e1746af3d1765dce681c9b99291d1bee