GS-6824: Extra backup due to socket timeout exception.
Tests at technion were conducted with InfiniBand  which uses backlog queue of 128. This causes connection requests to throw socket timeout if the requests can't be handled. When the GSM tries to instantiate a service on a new GSC, and socket timeout is thrown, it assumes that the service hasn't been instantiated and requests another instantiation. This causes an extra backup to be allocated.

Fix:
try and destroy the service that failed to instantiate.

svn path=/trunk/gigaspaces/; revision=55428


Former-commit-id: e3dab234a04ae88635ba25020a1246067ab9c614