GS-6927: Space copy from space to space using different security contexts.
Provide security context of remote space in internal API:

In new security model, we check permissions at the server and not at the proxy like we did before.
We need to check that the user has write permissions for the space being "copied to".
We need to check that the user has read permissions for the space being "copied from".
Note that it can be different users for both spaces.

Therefore, it is no longer enough to pass the security context of the space being "copied to".
We must also supply the security context of the space being "copied from". This way both checks
are applied equally at the server side.

svn path=/trunk/gigaspaces/; revision=56389


Former-commit-id: 93d2a61c8253142ddec32e7a205e299a8cbd5d33