GS-6962: A new log file is created for an undefined logger name.
If logger name doesn't exist in gs_logging.properties (like a logger defined in customer code) a new log file will be created.

Cause:
GSLogConfigLoader initializes the handler and assigns it to specific loggers, but the root logger hasn't been initialized. The LogManager initializes the root logger when an undefined logger is created - hence a new file is created.

Fix:
Assign the handlers to the root logger upon initialization.

svn path=/trunk/gigaspaces/; revision=56882


Former-commit-id: f2b7c5204b2935f9f6e01e970096a9ca7466ee22