GS-6972
Adding new built in processing unit container that simplifies the .NET user experience with processing unit.
The container name is BasicProcessingUnitContainer.
Adding new public class ProcessingUnitContainerHost that the user can use to programatically host processing unit containers for testing purposes.
(Does not integrate with service grid since it is not hosted in a gsc!!!)

svn path=/trunk/gigaspaces-dotnet/; revision=58918


Former-commit-id: 6536fcfd1d0cb57a1a6686d55b2db78e157c5c73