GS-6948
Changed exception thrown from Marshal(Input/Output)Stream due to class loader removal to be MarshalContextClearException. This exception is new and should never reach the network, it is transformed internally either to remote exception if occured at the client side or it causes the connection to be explicitly closed at the server side.

svn path=/trunk/gigaspaces/; revision=59386


Former-commit-id: ceb3105861b1eec3c01a8325fcde102959a4531a