GS-7208
Improving logical version infrastructure, adding stack of context because a single thread can handle an incoming context, which needs to be serialized with source platform logical version and then go to an outgoing invocation which should be handled with target platform logical version.
Moreover, an object when serialized can sometimes be serialized with a source context and sometimes with a target context.

svn path=/trunk/gigaspaces/; revision=62040


Former-commit-id: d79de9f81be2324f0e0a01d1850a858bb523b6cb