GS-7320
Adding IRedoLogFile interface which is used instead of linked list inside the memory redo log.
Implementing a MemoryRedoLogFile which is equivalent to the current redo log file and adding FixedSizeSwapRedoLogFile which uses external storage to swap packets once its memory size is reached.

Currently there is no real implementation of the IExternalRedoLogFileStorage which is the next stage.
Commiting unit tests of the entire project as well (TDD for the win)

svn path=/trunk/gigaspaces/; revision=62700


Former-commit-id: a6f33b62f3d929cbf7ac7f001617ce8e10a5b867