GS-7870: GSM fails to re-provision processing unit instance.
When cable is re-connected after service fault-detection handler has kicked in,
the no-longer monitored instance is added to the serviceBeanList via serviceAdded event from the LUS. 
This can eventually cause more service "instances" to be listed, and return a wrong result via the getActual method which counts the number of services in the list.

Fix -
getActual should return the number of monitored services, hence the number of service instances which have an active fault-detection handler monitoring them. This will avoid the distinction needed between services discovered and services provisioned.

svn path=/trunk/gigaspaces/; revision=69892


Former-commit-id: 39afdceeb628c048acd33cd93898f7227e02afc6