GS-7904: GSM should resolve an un-monitored service which is still pending to be provisioned.
After disconnecting the cable the service instance fault-detection kicks in, and it is placed in the pending to provision list until an available GSC is found. When re-connecting the cable, the service instance is discovered and added back via the LUS event. At this point, we can safely remove the instance from the pending list and open a fault-detection handler to start monitoring it.

Note that the "non-active" GSM already adds a fault-detection handler regardless!

svn path=/trunk/gigaspaces/; revision=70055


Former-commit-id: 102adef9c15c272ae595df40bb0e5578568dfdc0