GS-7942: GSM tries to instantiate an instance although previous shutdown is still in progress.
Added logging to see if an exception is thrown while trying to destroy service reported as failed from the FDH due to unhealthy state exception.

Unexport operation caused the destroy operation to break on a ClosedChannelException. The GSM caught the exception and continued to instantiate the failed service.

Fix - before the call to unexport, perform discard on GSC. 

svn path=/trunk/gigaspaces/; revision=70737


Former-commit-id: 0f889f5e281b2a3896fa45203ee041c30db57c04