GS-8033: Pending services at GSM may cause an extra-instance.
After GS-7904, GSM discovers services and removes them from the pending list if they have the same service-id.
But, if the service was re-instantiated (causing it to have a new service-id) while the GSM was disconnected, the GSM will not be able to remove it from the pending list. This will cause an extra instance once there is an available GSC.

Fix:
1. if service-id's don't match, try matching by presentation name.
2. Slow down execution of pending request when a new GSC is discovered - this is to allow discovery of available services.
3. Check if should dispatch pending request according to actual+pending<maintain 

svn path=/trunk/gigaspaces/; revision=71728


Former-commit-id: 2f53d64b2cb323dc56238f871b7bc8108373f69a