GS-8088
Add a connection monitor implementation that is based on the already existing failure detector of gigaspaces.
TBD rebuild failure detector to a monitor that is incharge of detecting dead connections and also detecting reconnections instead of each component in charge of detecting reconnection. This way we will be able to create full scenario tests simulating disconnection of not just replication but the entire space.

svn path=/trunk/gigaspaces/; revision=73915


Former-commit-id: c48b4b88f3b9973e48f9e40a2f77cdc7299e1714