GS-8088
Add the ability to configure consume timeout at sync replication.

We do not take todo_queue_timeout since it has different semantics at old replication. We have a different default for this (10 seconds) and we don't want to change the default for old replication due to different semantics.

todo-queue-timeout is not used in new replication.

svn path=/trunk/gigaspaces/; revision=80073


Former-commit-id: 6ae2b595e145de71006006f7f29ae24deaf14742