GS-8739: Not all processing unit instance incremental requests were fulfilled.
When performing multiple increment requests, some requests get ignored due to race condition between the instantiation and pending counts. the getActual() count already includes the incremented instance and the pending count also includes it since it was not yet removed.

The check (pending + actual < maintain) causes the request to be ignored (e.g. pending=3, actual=1, maintain=4 while the 'actual' counted instance is also a 'pending' instance).

When using an ESM which calls multiple increments it may cause it to block waiting for the instance to be discovered.

Fix: count pending instances and exclude the ones already provisioned

svn path=/trunk/gigaspaces/; revision=84350


Former-commit-id: f444a68aca02a28500bb305948dd7c9fde5ada62