GS-8777
Make sure to adjust the context class loader of a thread coming from .NET that tries to associate its AppDomain to the context class loader but it has no context class loader.

This occurs when creating an AppDomain explicitly in the .NET side with a new thread. This thread has no context class loader when it is attached to the JVM.

svn path=/trunk/gigaspaces/; revision=85258


Former-commit-id: a9e69a8b4e06a4248f412da08afaddd3af9ed2a7