GS-8831
Add to SpaceContext, ReplicationContext (for internal usage only) that right now indicates this operation came from a gateway without specifying the gateway name.

In future version we might want to extend this to support origin gateway name in order to be able to hold this data in space and also do decide on some custom replication filtering.

svn path=/trunk/gigaspaces/; revision=88667


Former-commit-id: 75a301ab6e876c2cc06fcdb0509b7145072ae49e