GS-8863, GS-6063
Removed SpaceModeAdaptiveMemoryManager usage, backup space will now throw memory shortage only when it is really in memory shortage state instead of always throwing it after it had reached memory shortage state once. In the new replication when memory shortage is thrown that replication packet is not confirmed and it will be treated as any other exception and retried until successful. As a result of this change, a backup will be relocated by the service only when it is constantly in memory shortage state and not when it was in that state momentarily.

svn path=/trunk/gigaspaces/; revision=90589


Former-commit-id: 8faa1406e1b5dc407083444e8533c4bf74118054