GS-9463
Restore previous behavior of replication router inside a space, it will be immediately available for replication once created instead of only after copy stage of recovery. This is because clean operation reconstruct the engine but do recovery process. We restore the behavior where we rely that the replication will not be connected during the copy stage because the space is registered at the LUS with replicatable=false during that stage.

svn path=/trunk/gigaspaces/; revision=98064


Former-commit-id: f5ed10b69e9fd73e7f8749de659156b51bcd4741