GS-9475
Adding special filter for the scenario where we have Central DB with LRU and it does not support versioning. In this case if we have a local view we need to send packets to the backup with previous server entry state so in case of fail over this packets can properly be matched against the templates in the view filter at the backup which became primary.
We do this because in this case the backup cannot reinsert the content to the packet because it does not hold any data

Reviewer: Anna

svn path=/trunk/gigaspaces/; revision=99762


Former-commit-id: a38122b5f89edbe6fef2cc8aa7126447385701cd