GS-10093
We have two built it filters, one for statistics and one for executor openspaces injection. 
This cause we create a full SpaceFilterEntry on each operation where the statistics do not require the object at all and the executor is only relevant for BeforeExecute. 

We can create an optimization that only create SpaceFilterEntry when needed and moreover if there is only single filter associated to the operation code we can skip the priority list invocation mechanism 

Reviewer: Dan

svn path=/xap/trunk/gigaspaces/; revision=115426


Former-commit-id: a7d3eae3719b788f32ee6b74c483b40a53157bdd