GS-10273
Added an AnnotatedCompressedMarshObjectConvertor that is used to convert from and to Compressed marsh object, but instead of using ObjectInputStream/ObjectOutputStream as done in  CompressedMarshObjectConvertor, it uses AnnotatedObjectOutputStream/AnnotatedObjectInputStream.
This AnnotatedObjectOutputStream/InputStream gets the right class loader in it's resolveClass method and it solved the bug (ClassNotFoundException) that occurred when used the ObjectInputStream/ObjectOutputStream.resolveClass.

Changed the convertor of BINARY storage type from MarshObjectConvertor to AnnotatedMarshObjectConvertor for the same reason.

Reviewer: Eitan.

svn path=/xap/trunk/gigaspaces/; revision=119622


Former-commit-id: 7f9cb52ce0d04ea0cc831037538959b385073681