GS-10467
A space will look for the target sink both via delegator and directly.
This is a baseline to allow connection between a space to the target gateway sink directly without the need for delegator. Which is a simplification requires for using gateways locally to synchronize clusters (or later on mirror) that does not need to go via delegator.

Reviewer: Dan

svn path=/xap/trunk/gigaspaces/; revision=125562


Former-commit-id: bae92aa28a91527453ade537643b029bf284b289