GS-10579
Add ClassLoaderThreadLocal which is aware to the context class loader and associate a seperate ThreadLocal map per class loader. Any get,set method is delegated to the corresponding ThreadLocal map for the context class loader.
This class uses the ClassLoaderCache and will self clean maps associated to class loaders which are removed implicitly or explicitly

Reviewer: Dank

svn path=/xap/trunk/gigaspaces/; revision=128327


Former-commit-id: 9520cfdcc8b8a3d495cf11056e2b1c7427980904