GS-10754
Before marking the current key that the copy stage is done, force a cycle of lease reaper which in its turn may create replication packets of entries that were expired during the copy stage. We need this in order for this packets to have an older key than the time of copy stage completion so the filter logic of the lease expired packets will know to filter it if they were not copied during the copy stage because their lease was already expired.

svn path=/xap/trunk/gigaspaces/; revision=131437


Former-commit-id: ea98b693f7ad2fe8f651a7185cbf717357b3c3c2