GS-10863
Do not call performReplication within removeTemplate scope of non notify template removal. We do not replicate this operation anyway but it causes the pending update operation on the replication context to be replicated while the entry physical lock is held instead of afterwards which causes replication sync mode deadlock with a transaction which has updated this object as well but due to fifo mode is waiting to acquire the same lock for some internal reasons.

svn path=/xap/trunk/gigaspaces/; revision=138067


Former-commit-id: 425f613c9cf5c3bd08b4e21f7d94a4aebcd9cd95