GS-12077 - LUS OOM in rare condition of plenty registration and unregistration of services due to disconnections
One of the LUS ran out of heap memory and proceeded to restart itself.
The shutdown appeared to be successful and it started a new LUS, but the LUS had issues communicating with our other LUS on another node.
Eventually the other node also ran into an issue with heap memory and followed the same path, resulting in having two new LUS deployed that would not communicate with each other.

Added limit to each registresion pending events size (500) after that the registration is cancled.
Test: LUSClietnHungTest.java

Reviewr: Moran.